migration method setup conversation
