+++ src/controllers/message.controller.js
@@ -26,6 +26,17 @@ class MessageController {
             })
         }).send({ response: res })
     }
+
+    static async setUpConversation(req, res, next) {
+		return new SuccessResponse({
+			code: 201,
+			message: "✔️ Conversation setup successfully",
+			metadata: await MessageService.setUpConversation({
+				...req.body,
+				userId: req.user._id,
+			}),
+		}).send({ response: res });
+	}
 }
 
 module.exports = MessageController;
