+++ src/services/message.services.js
@@ -2,6 +2,7 @@ const { FileNotFoundError, ForbiddenError } = require('../middlewares/error.resp
 const messageModel = require('../models/message.model');
 const { getReceiverSocketId, io } = require('../socket');
 const ConversationService = require('./conversation.services');
+const UserService = require('./user.services');
 
 class MessageService {
     static async sendMessage({ senderId, receiverId, message}) {
@@ -28,6 +29,25 @@ class MessageService {
         const conversation = await ConversationService.getConversation({participants});
         return conversation?.messages || [];
     }
+
+    static async setUpConversation({ email, userId }) {
+		const receiver = await UserService.searchUserByEmail({ email, userId });
+		const conversation = await ConversationService.setUpConversation({
+			senderId: userId,
+			receiverId: receiver._id
+		})
+		if (!conversation) throw new ForbiddenError("❌ Setup conversation error");
+		
+		// SEND SOCKET
+        const receiverSocketId = getReceiverSocketId(receiver._id);
+        const senderSocketId = getReceiverSocketId(userId);
+        if (receiverSocketId) {
+            io.to(receiverSocketId).emit("newConversation", conversation);
+            io.to(senderSocketId).emit("newConversation", conversation);
+        }
+
+		return conversation;
+	}
 }
 
 module.exports = MessageService;
